Spaces updates and lib effect refactor
I was having problems grokking the entirety of the different matrix listeners and functions, so I broke them up into different files.

Sadly I did this while also updating spaces code.

The current work on spaces is still incomplete at this point. If you sync a space hierarchy, then add a new room to a space it will not be reflected in the space children list, i will fix this tomorrow.

Issue #72