zion client refactor (#235)
Zion Client
mostly a "passthrough" abstraction that hides the underlying MatrixClient
normally, a shitty design pattern, but in zions case we want to
always encrypt
enforce space / channel relationships
get user wallet info
go to the blockchain when creating a space
go to the blockchain when updating power levels
etc
the zion client will wrap the underlying matrix client and
ensure correct zion protocol business logic
Key changes

we now have a zion client singleton, but the client in the context is only available when the matrix client is started, just like it was before
the tests and the app call the same code, adding a new route should be easier, adding side effects of routes should be easier
matrix has an extensive and complicated storage system (see the "live timeline") any duplication of this is going to be a disaster. I updated the use-my-profile hook to add event listers to matrix instead of piping data around and through zustand, i think this is the way that the room messages, and most things, are going to need to work eventually
Tests still pass, client-app and app both run normally locally

I still haven't touched the web3 login code, and plan to leave it alone. that stuff is pretty complicated