MetaMask hangs trying to get the gas fee for our contracts (#393)
MetaMask transaction for the ZionSpaceManager contract hangs. Looks like
MetaMask is repeatedly trying to get the gas fee estimate. Setting the
hardhat config to allowUnlimitedContractSize: true resolves this issue.
But this may be caused by our contract exceeding the size limit.

Filed issue here to track the proper fix:
https://github.com/HereNotThere/harmony/issues/392

This PR is a temporary workaround.

Co-authored-by: Tak Wai Wong <tak@hntlabs.com>