Separate hook to get space contract integration working (#399)
Created a separate hook to figure out how to get integrated space
creation, token entitlement working. Will refactor later to merge back
into the existing lib api.

In sample-app, click CREATE SPACE. 

It will create a space in Matrix, and a space in the contract. A bunch
of to-dos, and comments to remove code later once it is working. Still
missing - token entitlement configuration. Need to work with G on that.

Co-authored-by: Tak Wai Wong <tak@hntlabs.com>