wagmi integration (#473)
replace hand coded wallet integration with wagmi
left a few helpers in place, but we should be able to use all wagmi
natively


Issue #174, #250

The only weird thing i ran into, in order to have typed contracts in the
zion client, i
pass a provider into the zion client via the args parameter
I couldn't get the wagmi provider to work as an ethers provider. It's
possible I just wasn't
holding my mouth right, if I get into the internals of wagmi in the
future and figure it out
i'll update it.