Kerem/lint (#614)
- Introducing turborepo to handle monorepo-wide actions (like linting
every package in the workspace)
- Fix a quick linting error in a smart contract
- Introduce ci.yml: a github action that runs whenever we create a PR +
whenever we merge to main
- CI currently only runs `yarn lint` from the monorepo root, which runs
the `lint` command on every package listed under the yarn workspace -
and skips the package when the associated package.json doesn't have a
lint script
- `Warning:` had to remove the `install` script from
`matrix-zion-appservice`s package.json, because it kept auto-running a
command that kept failing on CI. I tried many approaches to make it work
but eventually timeboxed it for 2 hours and here we are. The script
doesn't run anything important, it merely checks for the existence of
some binaries. I deemed it safe to temporarily remove it, but calling it
out here now just in case it isn't. cc @tak-hntlabs