Integration test for Read permission (#680)
Added integration tests for the Read permission:
Test1: verifies that the token user is granted to join the space.
Test2: verifies that the non-token user is denied. But the test fails
because of this bug:
https://linear.app/hnt-labs/issue/HNT-205/createspacewithtokenentitlement-should-not-add-the-everyone-role-by

Added test util functions to facilitate the implementation of the
permission tests.

The test suite is currently disabled with "describe.skip". The test file
has comments about how to turn on gating check to run the test. Will
make this better in future PR.