Fix for losing messages, threads, and reactions after edits (#887)
Matrix assumes that all edits, reactions, and threads will be made
against the original message

When we launch, all “replace” event content is copied up to the
original.

Recursive replace events are not supported.

The Bug:
When we had a replace event come in while the app was running, we were
using the new id in the timeline, then any replies and reactions would
get lost on refresh, and any subsequent edits would get lost because of
the non recursive nature of replacments

The confusing part:
When we add a message locally, it generates a temporary id starting with
`~`. When the real id comes back from the server, we DO want to replace
this ID with the “official” one

The fix:
when a replace event comes in from the server, if it’s not a local event
replace, keep the old id

The other fix:
when i implemented this, the RichTextPreview stopped working, I had to
add a unique "key" property to the timeline to get this wot work. If you
add normal divs all the way down to lexical, things update fine. If you
click "edit" then "cancel" after you made an update, it works! But some
how the fast switching between submitting your edit and refreshing with
the next edit, somehow lexical doesn't pick it up.