unfurl links in messages (#912)
Enables unfurling in local development. Links in text will be unfurled
against mock content, of which there are 4 types:

- any image url (that ends in .jpg, .gif, etc) will return the same
image
- any giphy url (not an image file type extension) will return the same
giphy content
- any twitter link will return the same tweet
- any other link will return stack overflow content

Haven't tested this calling the live endpoint, but the mocks are all
actual responses.

There's bound to be lots of design/content issues around this in the
future, but these basic types should cover get us started.