Checking in the generated contract types (#976)
We get these changes after running `scripts/build-contract-types.sh`.

There will be a followup PR where if you don't check in the generated
types, it will block your merge. This way we can always know that the
contract build artifacts are kept in sync with dendrite. If it's merged
to main, it must be working with dendrite