HNT-489 Fix groups expanding (#1031)
Minor clean up + Potential fix: don't assume item's aren't to be
considered as "groups" based on the fact their height is zero but
instead on a bool. Issue being, resizeObserver only sets ref once, so if
the ref used to measure gets passed to the wrong element on first render
(based on the false assumption it's not a group) it will carry on
measuring from the wrong element. I guess this behaviour is valid.