token-worker: 'all' query (#1092)
Allows for passing an `all` query to the getTokens endpoint. Adding
`all` fetches all NFTs until pagination is exhausted. This is a hacky
workaround for alpha.

The frontend expects to be able to search for an NFT. Alpha was scoped
to have this be a client side fuzzy search. For users with a bunch of
NFTs, they're not going to be able to search them unless we load them
all at once.

In the client we could load more either by button click or scrolling,
which is what is currently in place. But that doesn't solve the issue,
as the user has to load more, search, load more, search, until all their
items have loaded.

For alpha we probably won't have anyone with a ton of NFTs on Goerli, so
we likely won't run into this issue at all - I imagine Alchemy's api
being called only once. In the long term, server and client will need to
be refactored.