Add foundryup command to the start local dev contraption (#1105)
going to start calling this a contraption

turns out if you don’t run foundryup you can get out of sync in an
unrecognizable way, and your contracts will build differently that they
do on the build machine

I added foundryup to the task and not to the start-local-blockchain.sh
file. I don’t know if this is right or wrong, it just seemed like the
right place to do it.