Update the spaceManagerContractHooks test (#1131)
It was always failing locally

I don’t know that this will really fix it, but it will make the output
less annoying…

Now that all of the tests create on chain spaces, the list is really
long and would scroll forever in the output window. This shortens it up
a bit.

Also i added a little extra time for the space to show up.