store transactions locally (#1150)
Idea is to store transactions when they are created, plus a listener
that deletes the transaction from storage once we have a receipt. The
user can navigate away or close their browser and will be informed of
success or failure once they log back in. Was testing with anvil block
time of 10 seconds.