basic auth header check for workers (#1186)
Speaking with @jterzis , the agreed on basic worker auth flow for reads
for alpha is:

1. store encoded secret in client during build, and include in header
2. check request headers in worker to verify

This PR adds a header check to the workers. A separate PR will add the
header in the client. Once these are done, we can manually store the
encoded secret in the Render dashboard for the client, as well in
Cloudflare for the workers.

Subsequently, CI work can be done to rotate the secret during
deployments.