show loading screen until wallet connection determined (#1245)
Adds a check in route logic to determine if wallet connection is
loading, and shows loading screen until it is. Without this check, a
user would see flashes of content for routes that they shouldn't: i.e. a
logged in user will see the Login button.
Once wallet connection has either connected or not, the rest of the
Routes logic can take over.