fix CreateSpaceListener has wrong abi (#1275)
Bug from V2 migration. The client is using the `eventsAbi` from lib's
`getSpaceFactoryEventsContractInfo()` which doesn't match the actual ABI
used by the space factory. Actually I can't find a generated ABI file
for the new space factory. I looked in `SpaceFactory.sol`, saw that it's
emitting the a new event `emit Events.SpaceCreated(_spaceAddress,
_msgSender(), spaceNetworkId);` (previously was CreateSpace), and just
hardcoded the ABI based off the event in
`spacesV2/libraries/Events.sol`.

This is the only place we're listening to events in the client
currently. I know some ABI stuff is being planned, so this solution
should be fine until it's sorted, but tag me if this ABI changes - the
tests mock this emitter so it's not gonna fail in CI