SpaceHome: fix loading states (#1253)
SpaceHome navigates users to a number of outcomes: the first channel,
the space threads page, or shows a join button.

This PR fixes:

Removes the space owner dependent logic to navigate to the
getting-started page. Instead the space owner will be directly navigated
there after creating space.

On fresh load there was an issue where users would be navigated to
threads page because the channels had not fully synced. I put in a hacky
fix for this.

On navigating between spaces, there was always a flash of the Join
button, even if you had already joined.