allow app to pass chain to load app and login (#1290)
This PR allows the app to pass a `chain` prop, which basically locks the
app to a specific network and ignores any network in the user's wallet
(or any network changes). This allows user to login to the app while on
another chain. We will still need guards in place for transactions
(prompt them to swap to goerli).

Originally started this issue to solve the app load error when on wrong
chain
https://linear.app/hnt-labs/issue/HNT-617/getting-white-page-if-im-not-on-mainnet-goreli-network-on-meta-mask,
but this PR does more than that.