added event handlers to zion client (#1324)
Context:
https://linear.app/hnt-labs/issue/HNT-609/as-hnt-i-want-to-know-ifhow-users-are-using-alpha#comment-4bdae9d6

Problem: We want to add user analytics. The best place to put them is
ZionClient, since it's the source of all important activity, in a way
that allows us to be DRY. However, we want to keep ZionClient as lean as
possible, while staying purely focused on business logic.

Idea: Add generalized event handlers to important functions. These
handlers can then be plugged in to be used for user analytics amongst
other things. By generalizing these into event handlers, we manage to
stay lean and focused on Business logic on Zion Client, while allowing
user analytics to be plugged into the center of all important user
activity.