Renounce ownership after space creation for space factory (#1378)
This PR Fixes an issue where the contract owner could make calls above Owner's
Permission. We initially decided to do this to let the SpaceFactory bootstrap the space creation by creating Owner, Everyone, and Extra Entitlement roles; however, the ownership of the contract was transferred to the SpaceOwner, allowing a space owner who has "sold" his NFT, to make Owner level updates still.

This change makes it so that the SpaceFactory renounces ownership of the Space contract and that the caller's permissions are checked at the function level.