CreateSpaceForm: token fetching improvements (#1402)
Prior iterations for token fetching relied on Alchemy API's paginated
results. Since then I updated our worker to return all the aggregated
results. So really we don't need the aggregation via react-query at all,
but I'm keeping the code in there for now in case this changes in the
future - if we were to switch to a load more button or scrolling
behavior

- fixes queryClient getting re instantiated
- fixes localhost testing behavior so `?vitalikTokens` query actually
works
- moves tracking selected tokens to store instead of react-hook-form
- simplifies getting cached token data
- bit better tests