integrate zion client event handlers (#1338)
context:
https://linear.app/hnt-labs/issue/HNT-609/as-hnt-i-want-to-know-ifhow-users-are-using-alpha#comment-4bdae9d6

This PR combines the ZionClient event handlers with Amplitude on the web
app, in the form of a Provider and a hook. The provider initializes
amplitude and prepares the proper useEffect hooks to update Zion
Client's event handlers as their dependencies update. For now, the
dependencies don't update, but they definitely will so it's good to have
this structure in place