track createspace (#1408)
of all 4 actions tracked, createSpace was the only one that didn't
appear on amplitude dashboard. the reason is that the web app has
multiple ways of creating a space (side note: we should consider
reducing it to a single way). I was able to find a function that all
different ways of creating a space eventually converged to, so i pulled
the event to that particular function `waitForCreateSpaceTransaction`