fix useUpdateChannelTransaction.test: Add room listeners to useChannelData so that it responds ot name changes (#1547)
The useChannelData returns a simplified version of the channel that is
all we need for rendering the left hand nav. When first built, I didn’t
consider channel name changes to be possible, much less account for
them.

Here is the channel data:
```
export interface Channel {
    id: RoomIdentifier
    label: string
    private?: boolean
    highlight?: boolean
}
```

It was previously populated from the result of the sync space
hierarchies call, which is run in useSyncSpaceHierarchies

This call is made once per space, and again whenever we detect a space
has new children via a EventType.SpaceChild event

This change adds a second stream of data to the useChannelData, and
pulls the name from the updated matrix room before returning it.

This fixes the useUpdateChannelTransaction, but leaves the list of
spaces stale. Ideally we would rebuild that list on any name change, in
the meantime, users can refresh their browser.