Fixes for getting extended key sharing running in the browser (#1544)
Now that things are live and presence is on I tried this out and nothing
worked. Fixes detailed below:

1) the local memory store doesn’t implement
`getSharedHistoryInboundGroupSessions` so that was never going to work
2) figured out a way to use indexdb in tests, so we can stop worrying
about that
3) updated the timeline item “key” so that the content will re-render
when it switches from “m.bad.encrypted” to m.message
4) the “currently_active” presence property seems broken in dendrite, i
was getting users that were “online” that had a 2 hour old last seen at,
so i updated my logic there.
5) cleaned up some logs
6) removed an extra decryptEventIfNeeded
7) removed the await empty promise after calling prepare to encrypt,
megaolm keeps an internal promise and I think? should wait for it there
(i don’t like it, but no cursing this time)
8) updated the max events per request. They ship out in reverse chrono
order, so this should be fine.
9) fixed a random bug in isRoomEncrypted, it doesn’t get wrapped in the
same try catch as everyone else, can throw on logout
10) added the wire content to the zion events, made it easier to debug
things
11) use importRoomKeys instead of fake dispatching a new room key event,
works better


TEST PLAN
open chrome and firefox
log in with different accounts
send some messages to a shared encrypted channel
in fire fox
- log out
- delete index db from the Storage section of developer tools
- log in
- wait for messages to show up