Kerem/fargate (#1534)
For dendrite's final evolution, we need to move away from ec2 instances
and use the fargate abstraction for ecs containers. Dendrite has
stateful dependencies on the filesystem, which we used to handle via
Elastic Block Store (EBS). But fargate doesn't support EBS, it only
supports EFS (Elastic File System). In order to move data into the EFS
file system, we create a bastion host too. Tl;dr:

1) Create bastion host
2) Create EFS
3) Create network rules to connect bastion and application networks to
EFS
4) Create fargate task definition that mounts to EFS
5) Create new ECS service that uses the new task definition

Remaining work: We need to configure the alb to work with the new
fargate service. Once we do that, we can decommission the ec2 instance,
the alb target group, the old ecs service and the old ecs task
definition. We will also have to reconfigure logz.io because it has
different rules for fargate