hnt-867 client app spinning in sync calls (#1591)
Based on logs, there are 104 'startClient' instances. Most likely, this
is caused by the useEffect in useZionClientListener trying to start the
client. More details in
[hnt-867](https://linear.app/hnt-labs/issue/HNT-867/client-app-spinning-in-sync-calls)

PR also fixed another issue hnt-857. If server is very slow like this
morning, the matrix client startClient( ) can time out with connection
error. We don't restart the matrix listener when that happens. Added
retry.

closes hnt-857, hnt-867

---------

Co-authored-by: Brian Meek <brian@hntlabs.com>