Fix the useaddRolChannel.test.tsx (#1629)
run singularly, this test usually passed, but running with the following
command it would always fail

```
cd clients/web/lib
yarn test --silent=0 roleGet.test.ts useAddRoleChannel.test.tsx
```

I identified two problems
- addRoleToChannelTransaction was getting called with an undefined role
id, so i added a waitFor that seems to have resolved that race
- the test was using a useRef to render the success value, so rendering
was left to chance. I switched this to a useState so that updating it
triggers a re-render

There is a third problem, we’re converting the role, which is a
BigNumber in the contract, to a number. If we end up with a lot of
roles, or if roles are given ids randomly instead of incrementally, our
app will break. We should probably keep our types consistent