Use expect(…).toHaveTextContent(…) instead of within(…).getByText(…) in tests (#1630)
It gives much better messages on failure. (it prints what’s there
instead and also renders the whole dom in most cases)

The only exception is when we want to assert that something is not in
the document, I left those.

There shouldn’t be any functional changes in this pr.