Fix flakyness in the roleDelete test (#1635)
The roleDelete test would fail for me almost every other time I ran all
tests.

Turns out that matrix doesn’t remove you from the room internally when
you call leave, i guess it’s waiting for the next sync, or or what, but
if it thinks you’re in the room and you try to join it again it will
just succeed and return the existing room.

I added waitFor assertions to ensure that we’re really out of the room
before trying to leave it again. This seems like it solves my
instability issues.