fix worker deps (#1643)
The workers weren't set up to run their tests in CI, and the updates in
#1636 broke a bunch of stuff in them. This PR:
- downgrades itty router
- upgrades siwe (siwe is different version between client & workers and
worker needs ver 2+)
- updates so workers run tests in CI
- skips the tests for siwe-worker. I checked out
https://github.com/HereNotThere/harmony/pull/1384 and the tests were
failing there too. It's because we import SpaceDapp in the worker and
I'm pretty sure the IStaticContractInfo `assert {type: json}` imports
break the tests but I don't know why. For now I've just left a comment
in the tests.