login screen ui (#1652)
- removed some of the SignupButtonStatuses as this appears to have
evolved since inception and it seems like they're not really doing
anything.
- removed tooltip that only showed errors on hover and added error
message below button
- styled the button a bit to be in line with Fei's updates

Notes: I messed around with having a bunch of metamask accounts and
toggling between them to see how the button was effected. Some
learnings:
- wagmi doesn't watch every account, only the connected ones
- if you have a connected account and switch to unconnected account,
wagmi doesn't pick that up, and it still signs you in as the first
connected account. It's up to user to see that they are signing in with
different account than the one they've selected in metamask
- if you have a connected account and you switch to another account,
wagmi updates to that account and we log you in with that account
- same behavior exhibited in rainbowkit and connectkit
- wagmi themselves say this is a rare use case (to have multiple
accounts in your wallet to switch between) and i think they're probably
right