HNT-916 Prevent message to send when selecting from typeahead menu (#1657)
Not ideal, but the PR submitted to lexical last year got reverted and I
don't feel I want to spend more time on this today. In a way the issue
is that the register hook within typeahead gets updated on every
keystroke too. Otherwise just changing the order of the JSX plugin tags
would have worked.