Kerem/infra refactor (#1631)
This PR moves the load balancer and the cluster out of the zion-node
module, onto the root of the environment. The reason we want this is so
we can add node2, node3, node4 etc under the production environment,
without having to create more load balancers. A single cluster will be
able to run multiple nodes moving forward. This is necessary for
consolidating logs, load balancers and deploys to ECS services.