HNT-948, HNT-951 Fix for mentions not showing up after login (#1689)
when we boot I wasn’t properly re-calculating events that came down in
the initial sync and then got decrypted. Especially if the user had to
use extended key sharing to fetch keys.

This change deletes some code that i was a little ashamed of, I was
calculating unread and mentions in two different places, once from
matrix events on initialization, and then from ztevents (zion events). I
was able to share the second bit of code and use it by waiting for the
initial event from the timeline store.

This code is classically “brittle” code. It's hard to read and hard to
modify, but it's working really well afik... hopefully we won't have to
touch it again until casablanca needs unread

Test Plan: included a new test that reproduced the bug