Update the Matrix room topic as a part of the 'useUpdateChannelTransaction' (#1695)
the channel description is an offchain field (aka m.room.topic). As part
of the useUpdateChannelTransaction flow, when the room topic is updated,
the client receives the matrix event, but doesn't update the MatrixRoom.
One of the reasons is that the MatrixRoom is an interface in the
matrix-js-sdk and it does not have a property for 'topic'. Hook up this
flow, and add 'topic' to our own Room type.