HNT-946 VList - Prevent endless layout shift loop (#1690)
The origin of the problem comes from Lexical preview component which
initialises taller and then shrinks after a tick (1). I still don't know
why that happens and would need to dig into the lexical internals to
figure this out. So part of the fix is to add a div around the preview
block to prevent this 😅 . This also removes a bunch of jitter that
occurred while scrolling.

The endless-loop occured when an item on the edge of the threshold
shrinks so it gets removed from the vlist, and added back again (2). To
prevent this we're now checking against the "historical" max height of
each item when deciding whether to include an item or not.

1. resizing issues before adding the wrapper


https://user-images.githubusercontent.com/557990/223455638-7e2f37ff-e0e8-4cab-9904-882173d234d7.mov

2. resizing loop:


https://user-images.githubusercontent.com/557990/223456190-3cb21045-b834-49f7-aa9e-ee5e13250c5f.mov

3. after fixes


https://user-images.githubusercontent.com/557990/223456409-9a64198f-a594-449a-b194-1c8f9fc0f5d5.mov