Refactor start local dev to wait for blockchain dependencies (#1725)
When we run DeployZionGovernance it outputs json files containing the
addresses of our smart contracts.
These addresses are checked in, and usually what's generated matches
whats checked in.
However, if someone forgets to check in the updated addresses after
changes to the smart contracts,
start local dev will end up in a bad state. We will deploy the updated
contracts to our local anvil
deployment, but the servers will launch before that completes and pick
up the old addresses. This has
tricked me more than once and is difficult to resolve, because you need
to stop everything
  
This change introduces stages to the start local dev, which will 

        // 1. yarn install
        // 2. start anvil
        // 3. deploy contracts
        // 4. starts all servers and clients

Now any local contract changes will be correctly referenced in local
server deployments