Kerem/hnt 987 pre work on terraform to finalize our (#1698)
Once we have multiple environments, it will be that much harder to make
changes to our generalized infra. This PR does a final round of
improvements before introducing the staging environment. Things like
moving the DB into the "node" hierarchy. Or removing references to
"zion" and replacing them with "dendrite"