change write authorization to use isEntitledToSpace check (#1681)
We perform a check when uploading an image on space owner. However, we
have a new permission `ModifySpaceSettings`, so potentially not only
owner should be able to upload an image. This PR updates the permission
check in siwe-worker to use `SpaceDapp.isEntitledToSpace(<Permission>)`.
Later we can expand it to account for .isEntitledToChannel if needed.
This is more flexible - if we want old behavior to check for owner, we
can just pass `Permission.Owner`.

- updates gateway-worker to pass along the permission to siwe-worker
- updates SpaceDapp check in siwe-worker to use entitlement check
- fixes siwe-worker tests
- refactors gateway-worker to allow for running local siwe-worker for
full authentication flow