operator nft token (#1731)
This PR provides a base ERC721 implementation along with ERC721A
optimization. It is intended for minting NFTs for a single wallet (or
for someone else). It also comes with signature minting using EIP-712,
and is a way for a contract admin to authorize an external party's
request to mint tokens on the admin's contract.

ELI5: You can authorize an external party to mint tokens on your
contract and specify what will be minted by that external party.

This ERC721 serves as the template for the Operator NFT, a Soulbound NFT
minted by a Node Manager contract when a user stakes their Token to the
network.