app: encrypted UI not showing up in threads (hnt-979) (#1788)
Digressed a tiny bit here, but this has been itching me for a while. 

The threads panel contain boxes with a `level2` background with on top
the` MessageTimeline` component that is also used in channels etc. The
problem this bug surfaces is that, the elements on the
`EncrpytionMessage` are of `level2` and blends with the background.

First solution was to add an absolute color with opacity (eg.
`rgba(255,255,255,0.1)`) - that would have a similar result on `level1`
as `level2` (i.e. a about 10% brighter). This works but adds yet another
token. Opacity only works for one layer, piling opacities isn't great
and hard to control.

The solution here is to add a prop (`elevate`) that can be used for
cards. It will bump one level, including for the child components.

```tsx
<Stack  elevate rounded="sm"> // displays as level2
    <Box padding background="level2">2</Box> // automatically bumps to level3
</Stack>
```

Like so: 

<img width="1552" alt="image"
src="https://user-images.githubusercontent.com/557990/225623428-c213c1ad-6996-4e71-b87b-b5496c15a95a.png">

This PR also adds background colors to `color` selection, esp. handy for
icons that should match background boxes etc. (`EncryptionMessage`) has
this key icon for instance.

<img width="709" alt="image"
src="https://user-images.githubusercontent.com/557990/225624316-17a44a86-17eb-4887-82fd-0d716ee00706.png">