Jt/hnt 609 amplitude worker (#1874)
Closes HNT-609

Prior to this PR, Amplitude was largely being blocked by ad-blockers
rendering it unusable for analytics tracking. This PR creates a simple
HTTP reverse proxy as a worker that runs on our zone towns.com. By doing
so, we avoid ad-blockers intercepting network requests from the client.

Currently the worker is deployed to test - amp-worker-test.towns.com

To do 
- [x] add env var `VITE_AMP_WORKER_URL` for amp-worker host to onrender.
- [x] establish a stable, pseudonymous identifier based on walletAddress
and set on wallet connect

To do after this PR is merged:
- [x] deploy workers to test, prod