token worker: add Infura and Alchemy compatible routes (#1894)
This adds new routes for 2 of the 3 endpoints we are using in the
client. For the third endpoint, we may go with
https://github.com/HereNotThere/harmony/pull/1879

This adds `getCollectionsForOwner/:provider/...` and
`:getCollectionMetadata/:provider/...` where `:provider` is either `al`
or `in`, pointing to Alchemy or Infura respectively, so the client can
switch easily. I'll do another PR to update the calls in the client,
then remove the old endpoints in the worker.

Closes HNT-1170
Closes HNT-1171