using shims instead of hardcoded nft address (#1879)
Tak and I were debugging something unrelated but noticed that there was
a hardcoded pioneer nft address. This PR attempts to remove that
address. Another thing we aim to accomplish here is to centralize
contract-specific business domain routines under the shims for
organization's sake. Finally, would like feedback on the removal of the
token-worker request in lieu of a balanceof check as @jterzis suggested
today

---------

Co-authored-by: Evan Salzbrenner <15717901+salzbrenner@users.noreply.github.com>