using env vars for node name (#1898)
This is one of the last stepping stones for CD. a more convenient way to
just pick an environment to deploy - without having to supply the node
name. We may change this in the future when a single environment needs
to support multiple nodes - which is unlikely for dendrite in the
foreseeable future.