app:  hook prototype for building absolute path (hnt-1088 hnt-1072) (#1915)
problem: link to profile is different depending on the context
(`/threads/profile`, `/channels/id/profile`,...) - initially this is a a
bit design problem. But components using the links aren't currently
aware of this (and shouldn't really have to). This is also related to
[hnt-1113](https://linear.app/hnt-labs/issue/HNT-1113/i-should-be-able-to-click-on-the-mention-and-see-the-users-profile)


1/ one way to solve this would be to handle the right-most panel as an
optional modal `channels/123/?infoPanel=profile&profileId=123`, that's
probably the most elegant way since we could reuse this for other
layouts and it would be an opt-in per view. However I haven't found a
way to make this play well with the router setup 🤔

2/ the most optimised, but also cumbersome/not very elegant would be to
generate a helper function or a base path and drill it down (via
context?) for each situation. eg: `const getProfileLink('channel',
{spaceId, channelId}); const getProfileLink('thread',{spaceId})`...

3/ or this. it works and could be extended, more clever if needed. I
somehow like a central way of handling links. Downside is that it feels
a bit heavy for what it does? The idea would be to be able to pass the
desired params (`profileId` here) and based on the current location it
would figure out the path...

anyways, I'm sure there's a more elegant / out-of-the-box way of doing
this, have the feeling using the data router in v6 would bring us closer
to that